0.1.6.1_Preview03 - preview release 03 for 0.1.6.1
Parts: Added new GoOnlineAndGoOfflineHandling.AcceptCustomChangeFromAttemptOnlineState option.  This new flag is only useful when combined with GoOnlineAndGoOfflineHandling.GoOnlineUpdatesBaseUseState.  When explicitly selected this new flag will check if the invoked PerformGoOnlineAction method explicitly changed the Base UseState from the caller set AttemptOnline value, and if so, tho outer method will not overwrite the invoke method's assigned value, as was previously the case and as is the case now when the new AcceptCustomChangeFromAttemptOnlineState option is not selected.  Modified BaseState IsUninitialized EM to optionally include (defaults to true) the OnlineUninitialized UseState.