0.1.6.1_Preview07 - preview release 07 for 0.1.6.1
E090: E090SubstInfo and E090SubstLocInfo: Added ToString variants to support removing the ObjID.FullName from the output when desired (it can be redundant in some cases).  Added IsDropRequested property as replacement for use of more error prone.  Modified Substrate Tracker DropRequestReason logic to request substrate removal as soon as the substrate is no longer AtWork and has reached a completed state, either with no job linked to it, or when the job itself has indicated it would like to be dropped.  Changed normal behavior so that once a job is linked to a substrate it remains linked to it even after the job has been dropped from other lists.  This allows the substrate to continue to see the job drop request reason.  Added trace log messages on substrate DropRequestReason changes and on JobTrackerLinkage instance changes.