0.1.6.1_Preview10 - preview release 10 for 0.1.6.1
Remoting: Refactored RemotingClient to support use of additional service commands (Connect, Disconnect, GoOnline, GoOnlineAndInitialize, GoOffline).  Resolved issue with SessionManager where reception of Ack-only buffer at the session level could trigger a null reference exception (only relevant for datagram transport types).  Changes to PatchPanel transport to improve use with a debugger.

Part: Changed description and implementation of the SimpleActivePartBehaviorOptions.MainThreadSetsStateToOffline behavior so that it only does this if state is Undefined or Initial at the start of the main thread.  Otherwise it infers that earlier code explicitly set the UseState and and thus does not change the state in the base class at main thread start.

E090: added comment about impending change to ITransferPermissionRequest interface that is expected to arrive as part of 0.1.6.2.

Remoting Tool: change per addition of support for new service commands as described above.