0.1.7.0_PR01 Preview 01
Initial preview release for 0.1.7.0.  This is a preview release and gives a preview of some of the new capabilities that will be in the final version of 0.1.7.0.  The detailed implementation for these new capabilities are not entirely complete and are still subject to change prior to the final version of this release.

This preview includes significant changes in the E090 Substrate Scheduling related class tree including breaking changes both the public and internal APIs that are used in this class tree.

E090: SubstrateRouting: Added Comment to SubstrateRoutingItemBase which client can use to improve source traceability of for individual SRM items and sets thereof in logs.  Changed ITransferPermissionRequest through the addition of a ITransferPermissionState, TransferPermissionSummaryStateCode and two default state implementation classes: TransferPermissionState and TransferPermissionStateForPublication.  This is a breaking change as the prior ITPR state publisher property only carried a token set.  This has now been replaced with an ITransferPermissionState which caries the related token set in it in addition to a number of other related properties.  The primary functional difference is the addition of the TransferPermissionSummaryStateCode which allows a client to check if a given location handler part is NotAvailable (transfer request will fail), Busy (transfer request will block and then likely succeed), Available (transfer request will succeed immediately), or AlmostAvailable (transfer request is predicted to succeed after a relatively short period of time).  In addition the ISubstrateRoutingManager interface now requires support for a RetractArmsAndReleaseAll action factory method.  The previous ISRM.DetailsNVSPublisher property has been removed.

E090: SubstrateScheduling: All scheduling related interfaces and classes have been modified/extended including the part<->tool interface in order to support scheduler behavior selection controls (individual enables for service, recovery and automatic processing related behaviors) which can be changed both externally and internally, typically as requested by the tool in relation to a selected annunciator (recovery) action.  The scheduler part is expected to implement the ISubstrateSchedulerPart interface which also requires that it publish an ISubstrateSchedulerPartState.  This state is also the basis for the information that the part passes to the tool to indicate the overall state and to inform the tool of what behaviors are currently enabled.  There are two default implementations for this required state: SubstrateSchedulerPartState and SubstrateSchedulerPartStateForPublication.
The tool now uses a new RequestsAndStatusOutFromTool which allows it to give both status information to the hosting part (ToolFaultReason for example) and carries the prior RequestNVSFromTool.  Some process specific parts may now implement a new IPrepare interface which allows the scheduler part/tool to observe their published IPreparednessState (with two default implementation classes) and to request that the part prepare for a given processSpec and processStepSpec.  ISubstrateTrackerBase now supports a separate concept of IsServiceNeeded from its prior IsUpdateNeeded.  This supports consolidation of the concepts of updating the tracked state of a corresponding substrate and having external control over when any "service" specific recalculation logic is triggered.  Changed the SubstrateStateTally from a struct to a class and changed its corresponding internal use model (primarily addition of Clear method).

E090 SubstrateTestingTools: The example classes provided here have been modified to make use of the new interfaces and general capabilities described above.  The test part now supports external requests to change the currently selected behavior both explicitly and as an optional part of a GoOnline(true) action.  The test tool now supports use of an Error IANSource which is used to report SRM action failures and to solicit guidance on how to address them (Continue, Return, Hold).  It supports a similar prepare failed Error IANSource with Retry, AbortJob, StopJob and PauseJob options.  It makes use of a new evaluation mechanism that allows it to recognize more complex cross connections in routing and processing options.  It makes use of this new evaluation mechanism combined with the new ITPR AlmostAvailable concept in order to get and pre-position a substrate in front of the next process location slightly before the location becomes ready to have the transfer be requested and performed.  This new logic is used to support swap operations with dual arm robots.  The supporting test process modules have been modified to support the new ITPR interface and the SimpleExampleProcessModuleEngine also supports the IPrepare interface.

Modular.Action: Added AddNotifyOnCompleteInline EM to simplify call chain style building of modular actions.

Modular.Common: Adding early version of ValueContainer.ConvertToRawJSON EM which is used to produce JSON strings from supported ValueContainer contents that attempt to represent just the value contents without unnecessary type wrappers.  In particular this method is intended to support conversion of NVS objects to JSON dictionary format [{"key":value},...].

Modular.Config: ConfigIVIBridge - modified Sync behavior to force traversal of the lookAtLater list.

Modular.Interconnect/WCF.cs - this source file has been removed as the functionality intended here is not fully implemented through Modular.Interconnect.Remoting.

Modular.Part: Added ExplicitFaultReason to part BaseState (et. al.) which can be used to report the presence of a potentially resolvable fault condition without needing to change the current UseState.  This is generally intended to be used when a part has separately posted a fault indication using some mechanism like that supported via an ANManager.

Utils.Collections.TokenSet is now DataContract serializable.

Utils.Disposable: Changed InvokeDelegateOnDispose to be a sealed class and changed it to be derived from DisposableBaseBase in place of prior DisposableBase.  This removes the need to interact with the finalization mechanisms which this helper class did not support anyway.

Utils.Pooling:  Added BasicFreeList.ReleaseAndClear(list) and BasicFreeList.ReleaseSet(set) helper methods.

PartsLib LPM: changed ILPMPart to require support for the E090.SubstrateRouting.ITransferPermissionRequest interface.

PartsLib MDRF: added new MDRFFileReaderBehavior.AttemptToDecodeOccurrenceBodyAsNVS selectable reader behavior.

WPF E090View: small change to render locations that have a non-empty NotAccessibleReason using grey/darkening in place of prior error color (sienna).

WPF: IndexIntoSplitStringConverter now supports a MapNullOrEmptyStringTo enumeration to determine what value is returned when there is no token at the indexed position.  Added IndexIntoSplitStringConverter derived variants for '|', ':', and ';'